Rename workspace: voicers -> voice-*
- voicers     -> voice-tts   (core TTS library)
- voicers-cli -> voice-cli   (binary installs as 'voice')
- voicers-dsp -> voice-dsp   (DSP primitives)
- voicers-g2p -> voice-g2p   (grapheme-to-phoneme)
- voicers-nn  -> voice-nn    (neural network modules)

The namespace supports future crates: voice-stt (Whisper),
voice-stream, etc. The CLI binary is named 'voice' so users
can run: voice --text "Hello world" --play
