+++ crates/voice-cli/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "voice-cli"
+version = "0.1.0"
+edition = "2021"
+rust-version = "1.85.0"
+description = "CLI for voice TTS — like say, but with Kokoro"
+
+[[bin]]
+name = "voice"
+path = "src/main.rs"
+
+[dependencies]
+voice-tts = { path = "../voice-tts" }
+voice-g2p = { path = "../voice-g2p" }
+clap = { version = "4", features = ["derive"] }
+rodio = "0.20"
+hound = "3"
