+++ crates/voice-g2p/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "voicers-g2p"
+name = "voice-g2p"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.85.0"
-description = "Grapheme-to-phoneme conversion for voicers"
+description = "Grapheme-to-phoneme conversion: misaki dictionary + espeak-ng fallback"
 
 [dependencies]
 fancy-regex = "0.14"
