+++ crates/voice-nn/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "voicers-nn"
+name = "voice-nn"
 version = "0.1.0"
 edition = "2021"
 rust-version = "1.85.0"
-description = "Neural network modules for voicers: ALBERT, BiLSTM, vocoder, prosody"
+description = "Neural network modules: ALBERT, BiLSTM, vocoder, prosody"
 
 [dependencies]
 mlx-rs = { version = "0.25" }
 mlx-macros = "0.25.3"
 serde = { version = "1", features = ["derive"] }
-voicers-dsp = { path = "../voicers-dsp" }
+voice-dsp = { path = "../voice-dsp" }
