+++ crates/voice-nn/src/lstm.rs
