+++ crates/voice-nn/src/text_encoder.rs
