+++ crates/voice-tts/src/error.rs
