--- crates/voicers-cli/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "voicers-cli"
-version = "0.1.0"
-edition = "2021"
-rust-version = "1.85.0"
-description = "CLI for voicers TTS"
-
-[dependencies]
-voicers = { path = "../voicers" }
-voicers-g2p = { path = "../voicers-g2p" }
-clap = { version = "4", features = ["derive"] }
-rodio = "0.20"
-hound = "3"
